Removed the intellisense files, and references the definitions instead.
