+++ package.json
@@ -25,7 +25,7 @@
   "homepage": "https://gunjandatta.github.io",
   "dependencies": {
     "core-js": "^2.6.0",
-    "gd-sprest-def": "^0.4.4"
+    "gd-sprest-def": "^0.4.5"
   },
   "devDependencies": {
     "@babel/core": "^7.2.0",
