--- src/intellisense/entityData.d.ts
@@ -1,34 +1,0 @@
-/**
- * Entity Data
- */
-export interface IEntityData {
-    /** Account Name */
-    AccountName?: string;
-
-    /** Department. */
-    Department?: string;
-
-    /** EMail */
-    Email?: string;
-
-    /** */
-    IsAltSecIdPresent?: boolean;
-
-    /** Mobile Phone */
-    MobilePhone?: string;
-
-    /** Object ID */
-    ObjectId?: string;
-
-    /** Principal Type */
-    PrincipalType?: string;
-
-    /** SharePoint Group ID */
-    SPGroupID?: string;
-
-    /** SharePoint User ID */
-    SPUserID?: string;
-
-    /** Title */
-    Title?: string;
-}
