+++ src/intellisense/index.d.ts
@@ -1,17 +1,3 @@
-export * from "./entityData";
 export * from "./graph";
-export * from "./navigation";
-export * from "./odata";
-export * from "./peopleManager";
 export * from "./peoplePicker";
-export * from "./profileLoader";
-export * from "./propertyValues";
-import * as Results from "./results";
-export * from "./search";
-export * from "./socialFeed";
-export * from "./tenantApp";
-export * from "./tenantAppCatalog";
-export * from "./tenantApps";
-export * from "./userProfile";
 export * from "./utility";
-export { Results }
