--- src/intellisense/results.d.ts
@@ -1,218 +1,0 @@
-import { SP } from "gd-sprest-def";
-
-/**
- * Results
- */
-export interface IResults<P> {
-    /** The collection data */
-    results: Array<P>;
-}
-
-/**
- * App License
- */
-export interface AppLicense {
-    /** The collection of app licenses. */
-    results: Array<SP.AppLicense>;
-}
-
-/**
- * Content Type Id
- */
-export interface ContentTypeId {
-    /** The collection of content type ids. */
-    results: Array<SP.ContentTypeId>;
-}
-
-/**
- * Creatable Item Information
- */
-export interface CreatableItemInfo {
-    /** The collection of creatable item information. */
-    results: Array<SP.CreatableItemInfoCollection>;
-}
-
-/**
- * Custom Action Element
- */
-export interface CustomActionElement {
-    /** The collection of custom action elements. */
-    results: Array<SP.CustomActionElement>;
-}
-
-/**
- * Custom Result
- */
-export interface CustomResult {
-    /** The collection of custom results. */
-    results: Array<CustomResult>;
-}
-
-/**
- * Data Row
- */
-export interface DataRow {
-    /** The collection of data rows. */
-    results: Array<SP.SimpleDataRow>;
-}
-
-/**
- * Key/Value
- */
-export interface KeyValue {
-    /** The collection of key-value pairs. */
-    results: Array<SP.KeyValue>;
-}
-
-/**
- * List Data Validation Failure
- */
-export interface ListDataValidationFailure {
-    /** The collection of list data validation failures. */
-    results: Array<SP.ListDataValidationFailure>;
-}
-
-/**
- * Menu Node
- */
-export interface MenuNode {
-    /** The collection of menu nodes. */
-    results: Array<SP.MenuNode>;
-}
-
-/**
- * Number
- */
-export interface Number {
-    /** The collection of numbers. */
-    results: Array<number>;
-}
-
-/**
- * Object
- */
-export interface Object {
-    /** A collection of objects. */
-    results: Array<object>;
-}
-
-/**
- * Parameter Information
- */
-export interface ParameterInformation {
-    /** The collection of parameter information. */
-    results: Array<SP.ParameterInformation>;
-}
-
-/**
- * Personal Result Suggestion
- */
-export interface PersonalResultSuggestion {
-    /** The collection of personal result suggestions. */
-    results: Array<PersonalResultSuggestion>;
-}
-
-/**
- * Query Result
- */
-export interface QueryResult {
-    /** The collection of custom results. */
-    results: Array<QueryResult>;
-}
-
-/**
- * Query Suggestion Query
- */
-export interface QuerySuggestionQuery {
-    /** The collection of query suggestion queries. */
-    results: Array<QuerySuggestionQuery>;
-}
-
-/**
- * Refiner
- */
-export interface Refiner {
-    /** The collection of refiners. */
-    results: Array<Refiner>;
-}
-
-/**
- * Refiner Entry
- */
-export interface RefinerEntry {
-    /** The collection of refiner entries. */
-    results: Array<RefinerEntry>;
-}
-
-/**
- * Shared User
- */
-export interface SharedWithUser {
-    /** The collection of shared users. */
-    results: Array<SP.SharedWithUser>
-}
-
-/**
- * Social Actor
- */
-export interface SocialActor {
-    /** The collection of social actors. */
-    results: Array<SocialActor>;
-}
-
-/**
- * Social Data Item
- */
-export interface SocialDataItem {
-    /** The collection of social data items. */
-    results: Array<SocialDataItem>;
-}
-
-/**
- * Social Data Overlay
- */
-export interface SocialDataOverlay {
-    /** The collection of social data overlays. */
-    results: Array<SocialDataOverlay>;
-}
-
-
-/**
- * Social Post
- */
-export interface SocialPost {
-    /** The collection of social posts. */
-    results: Array<SocialPost>;
-}
-
-/**
- * Social Post Definition Data Item
- */
-export interface SocialPostDefinitionDataItem {
-    /** The collection of social post definition data items. */
-    results: Array<SocialPostDefinitionDataItem>;
-}
-
-/**
- * Social Thread
- */
-export interface SocialThread {
-    /** The collection of social threads. */
-    results: Array<SocialThread>;
-}
-
-/**
- * String Collection
- */
-export interface String {
-    /** The collection of strings. */
-    results: Array<string>;
-}
-
-/**
- * Visualization Field
- */
-export interface VisualizationField {
-    /** The collection of visualization fields. */
-    results: Array<SP.VisualizationField>;
-}
